+++ plugins/postcss-fit-tolerance-property/src/index.ts
@@ -1,4 +1,4 @@
-import type { PluginCreator } from 'postcss';
+import type { Declaration, PluginCreator } from 'postcss';
 
 /** postcss-fit-tolerance-property plugin options */
 export type pluginOptions = {
@@ -6,7 +6,8 @@ export type pluginOptions = {
 	preserve?: boolean,
 };
 
-const FIT_TOLERANCE_REGEX = /^fit-tolerance$/i;
+const IS_FIT_TOLERANCE_REGEX = /^fit-tolerance$/i;
+const IS_FLOW_TOLERANCE_REGEX = /^flow-tolerance$/i;
 
 const creator: PluginCreator<pluginOptions> = (opts?: pluginOptions) => {
 	const options: pluginOptions = Object.assign(
@@ -21,9 +22,15 @@ const creator: PluginCreator<pluginOptions> = (opts?: pluginOptions) => {
 	return {
 		postcssPlugin: 'postcss-fit-tolerance-property',
 		Declaration(decl): void {
-			if (!FIT_TOLERANCE_REGEX.test(decl.prop)) {
+			if (!IS_FIT_TOLERANCE_REGEX.test(decl.prop)) {
 				return;
 			}
+
+			if (hasFallback(decl)) {
+				return;
+			}
+
+			// TODO: check for fallback
 
 			// Insert the new value before the current value.
 			decl.cloneBefore({
@@ -40,6 +47,25 @@ const creator: PluginCreator<pluginOptions> = (opts?: pluginOptions) => {
 		},
 	};
 };
+
+function hasFallback(node: Declaration): boolean {
+	const parent = node.parent;
+	if (!parent) {
+		return false;
+	}
+
+	for (const sibling of parent.nodes) {
+		if (sibling === node) {
+			continue;
+		}
+
+		if (sibling.type === 'decl' && IS_FLOW_TOLERANCE_REGEX.test(sibling.prop)) {
+			return true;
+		}
+	}
+
+	return false;
+}
 
 creator.postcss = true;