+++ plugins/postcss-private-rule/test/_tape.mjs
@@ -5,6 +5,12 @@ postcssTape(plugin)({
 	basic: {
 		message: 'supports basic usage',
 	},
+	nesting: {
+		message: 'supports complex nested rules',
+	},
+	encoding: {
+		message: 'all encoded values are valid css',
+	},
 	'examples/example': {
 		message: 'minimal example',
 	},