+++ plugins/postcss-private-rule/CHANGELOG.md
@@ -3,6 +3,8 @@
 ### Unreleased (patch)
 
 - Only assign new property names to the PostCSS AST when the name actually changed
+- Fix private property lookup for nested rules
+- Fix lexical scoping
 
 ### 1.0.0